+++ model/exercise_association.go
@@ -9,9 +9,9 @@ import (
 )
 
 type ExerciseAssociation struct {
-	ID         uint `gorm:"primary_key"`
-	ExerciseID uint
-	TableID    uint
+	ID         int64 `gorm:"primary_key"`
+	ExerciseID int64
+	TableID    int64
 	CreatedAt  time.Time
 	UpdatedAt  time.Time
 }
@@ -32,7 +32,7 @@ func NewExerciseAssociationFlow() *ExerciseAssociationFlow {
 }
 
 // InsertExerciseAssociation 将exerciseID与tableIDList的关系插入表中
-func (*ExerciseAssociationFlow) InsertExerciseAssociation(exerciseID uint, tableIDList []uint) error {
+func (*ExerciseAssociationFlow) InsertExerciseAssociation(exerciseID int64, tableIDList []int64) error {
 	for _, tableID := range tableIDList {
 		exerciseAssociationDAO := &ExerciseAssociation{ExerciseID: exerciseID, TableID: tableID}
 		if err := GetSysDB().Transaction(func(tx *gorm.DB) error {
